+++ OpenSystemDynamics/src/settings.js
@@ -2,8 +2,6 @@ var Settings = {
 	fileExtension: ".ssd",
 	// Maximum recent files to be displayed in menu
 	MaxRecentFiles: 10,
-	// Controls the zoom in NW mode
-	nwInitZoom: 0.5,
 	// The font size under the primitives
 	primitiveFontSize: 13,
 	// Whatever to show debug