+++ distribute/package-for-mac/url-config.sh
@@ -1,4 +1,0 @@
-# This file is intended to be sourced from other files
-
-# Find latest version on https://dl.nwjs.io/
-NWJS_URL="https://dl.nwjs.io/v0.115.0/nwjs-sdk-v0.115.0-osx-arm64.zip"