+++ distribute/package-for-win32/download.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-source ../nwjs-url-config.sh
-../helper-scripts/download-nwjs.sh win32 $NWJS_WIN32_URL