+++ distribute/package-for-win32/package.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-source ../nwjs-url-config.sh
-../helper-scripts/package-for-win-helper.sh win32 $NWJS_WIN32_URL