+++ distribute/package-for-win64/download.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-source ../nwjs-url-config.sh
-../helper-scripts/download-nwjs.sh win64 $NWJS_WIN64_URL