+++ distribute/package-for-win64/package.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-source ../nwjs-url-config.sh
-../helper-scripts/package-for-win-helper.sh win64 $NWJS_WIN64_URL